remove outdated comment and add check to autostart login

+++ src/manifest.sh
@@ -257,8 +257,6 @@ distrobox_create_pod() {
         flags=($argFlag "$(sanitizeString "$argValue")" "${flags[@]}")
     done
     
-    # passing split args required here to ensure distrobox picks them up as individual flags and not one massive flag
-    # shellcheck disable=SC2086
     distrobox create "${flags[@]}"
 }
 
@@ -476,7 +474,9 @@ if [[ -n "${__CONTAINER_AUTOSTART_AFTER_BUILD+x}" ]] && [[ "${__CONTAINER_AUTOST
 fi
 
 if [[ -n "${__CONTAINER_AUTOSTART_LOGIN+x}" ]] && [[ "${__CONTAINER_AUTOSTART_LOGIN}" -eq "true" ]]; then
-    cp "$HOME/.local/share/applications/$__CONTAINER_NAME.desktop" "$HOME/.config/autostart/"
+    distrobox-generate-entry "$__CONTAINER_NAME)"
+    
+    [[ -f "$HOME/.config/autostart/$__CONTAINER_NAME.desktop" ]] || cp "$HOME/.local/share/applications/$__CONTAINER_NAME.desktop" "$HOME/.config/autostart/"
 fi
 
 exit 0
